User service and controller refactor, removed Transactional annotations, temporary commented loggin in other controller

+++ src/main/java/org/dev/cash_accounts_manager_backend/dtos/PagedResponse.java
@@ -9,9 +9,9 @@ import java.util.List;
  */
 public record PagedResponse<T>(
         List<T> content,
-        int pageNumber,
-        int pageSize,
-        int pagesCount,
-        int currentPageElementsCount,
-        int totalElementsCount
+        long pageNumber,
+        long pageSize,
+        long pagesCount,
+        long currentPageElementsCount,
+        long totalElementsCount
 ) { }
